[Mise à jour] ajout de la page pour afficher les cartes de tous les pokémons, ajout des composants pour chaque section et ajout des composants pour le Header et le Footer , amélioration des structures des codes
