+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import LandingPage from "./compenents/LandingPage"
-import PokemonCard from "./compenents/PokemonCard"
+import LandingPage from "./pages/LandingPage"
+import PokemonPage from './pages/PokemonPage';
 
 function App() {
  
@@ -10,7 +10,7 @@ function App() {
     <Router>
       <Routes>
         <Route path='/' element={<LandingPage/>}/>
-        <Route path='/pokemon' element={<PokemonCard/>}/>
+        <Route path='/pokemon' element={<PokemonPage/>}/>
         <Route></Route>
       </Routes>
     </Router>
